Refactor the rest of the apps to use component containers
