+++ yahoo-fluxible/js/components/CartContainer.jsx
@@ -1,6 +1,7 @@
 'use strict';
 
 var React = require('react/addons');
+var Cart = require('../../../common/components/Cart.jsx');
 var StoreMixin = require('fluxible').StoreMixin;
 var CartStore = require('../stores/CartStore');
 var cartCheckout = require('../actions/cartCheckout');
@@ -11,7 +12,7 @@ var Product = React.createClass({
     }
 });
 
-var Cart = React.createClass({
+var CartContainer = React.createClass({
     mixins: [StoreMixin],
 
     statics: {
@@ -31,7 +32,7 @@ var Cart = React.createClass({
         return this._getStateFromStores();
     },
 
-    checkout: function () {
+    onCheckoutClicked: function () {
         if (!this.state.products.length) {
             return;
         }
@@ -41,26 +42,8 @@ var Cart = React.createClass({
     },
 
     render: function () {
-        var products = this.state.products;
-
-        var hasProducts = products.length > 0;
-        var nodes = !hasProducts ?
-            <div>Please add some products to cart.</div> :
-            products.map(function (p) {
-                return <Product key={p.id}>{p.title} - &euro;{p.price} x {p.quantity}</Product>;
-            });
-
         return (
-            <div className="cart uk-panel uk-panel-box uk-panel-box-primary">
-                <div className="uk-badge uk-margin-bottom">Your Cart</div>
-                <div className="uk-margin-small-bottom">{nodes}</div>
-                <div className="uk-margin-small-bottom">Total: &euro;{this.state.total}</div>
-                <button className="uk-button uk-button-large uk-button-success uk-align-right"
-                    onClick={this.checkout}
-                    disabled={hasProducts ? '' : 'disabled'}>
-                    Checkout
-                </button>
-            </div>
+            <Cart products={this.state.products} total={this.state.total} onCheckoutClicked={this.onCheckoutClicked} />
         );
     },
 
@@ -69,4 +52,4 @@ var Cart = React.createClass({
     }
 });
 
-module.exports = Cart;
+module.exports = CartContainer;
